switch from reqwest and q to fetch and promises
