+++ public/js/services/AtomsApi.js
@@ -1,21 +1,29 @@
-import { pandaReqwest } from './pandaReqwest';
+import { pandaFetch } from './pandaFetch';
 import { getStore } from '../util/storeAccessor';
 
 export default {
 
   fetchAtom: (atomId, atomType) => {
-    return pandaReqwest({
-      url: `/api/preview/${atomType}/${atomId}`,
-      method: 'get'
-    });
+    return pandaFetch(
+      new Request(
+        `/api/preview/${atomType}/${atomId}`,
+        {
+          method: 'get'
+        }
+      )
+    );
   },
 
 
   createAtom: (atomType) => {
-    return pandaReqwest({
-      url: `/api/preview/${atomType}`,
-      method: 'post'
-    })
+    return pandaFetch(
+      new Request(
+        `/api/preview/${atomType}`,
+        {
+          method: 'post'
+        }
+      )
+    )
   }
 
 }
